requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
bidict==0.22.1
blinker==1.6.3
bottle==0.12.25
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dnspython==2.4.2
EasyProcess==1.1
entrypoint2==1.1
eventlet==0.33.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mediapipe==0.10.7
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
mysql-connector-python==8.1.0
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
protobuf==3.20.3
proxy-tools==0.1.0
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pygame==2.5.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Frameless-Window==0.2.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytweening==1.0.7
pywin32==306
redis==5.0.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
sv-ttk==2.5.5
urllib3==2.0.6
websockets==11.0.3
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
